+++ R/graph_functions.R
@@ -70,7 +70,6 @@
 #'
 #' event_line(res, spread = 100, start_date = "2010-12-01",
 #' end_date = "2011-06-30", category = TRUE)
-#'
 #' }
 event_line <- function(data,
                        x = t,